pyproject.toml
[tool.poetry]
name = "lupinsummary"
version = "0.1.0"
description = "Creates summaries of clinical reports"
authors = ["Kartik-007 <masterkartik1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9, <3.9.7"
openai = "^1.17.1"
langchain = "^0.1.16"
sentence-transformers = "^2.6.1"
chromadb = "^0.4.24"
pymupdf = "^1.24.1"
unstructured = "^0.13.2"
streamlit = "^1.33.0"
python-dotenv = "^1.0.1"
pypdf2 = "^3.0.1"
fpdf = "^1.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
